Revert "[frontend] Refactor bilan page"
