--- frontend/src/components/bilans/DeleteDialog.tsx
@@ -1,42 +1,0 @@
-'use client';
-
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
-
-interface DeleteDialogProps {
-  open: boolean;
-  onCancel: () => void;
-  onConfirm: () => void | Promise<void>;
-}
-
-export default function DeleteDialog({
-  open,
-  onCancel,
-  onConfirm,
-}: DeleteDialogProps) {
-  return (
-    <AlertDialog open={open} onOpenChange={onCancel}>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>Supprimer ce bilan ?</AlertDialogTitle>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>Annuler</AlertDialogCancel>
-          <AlertDialogAction
-            className="bg-red-600 hover:bg-red-700"
-            onClick={onConfirm}
-          >
-            Supprimer
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
