+++ frontend/src/pages/MesBilans.tsx
@@ -1,18 +1,37 @@
 'use client';
 
-import { Plus } from 'lucide-react';
+import {
+  FileText,
+  Plus,
+  ChevronLeft,
+  ChevronRight,
+  Trash2,
+} from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
-import EmptyState from '@/components/bilans/EmptyState';
-import BilansTable from '@/components/bilans/BilansTable';
-import PaginationControls from '@/components/bilans/PaginationControls';
-import DeleteDialog from '@/components/bilans/DeleteDialog';
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog';
 import { NewPatientModal } from '@/components/ui/new-patient-modal';
 import { ExistingPatientModal } from '@/components/ui/existing-patient-modal';
 import { useAuth } from '../store/auth';
@@ -66,6 +85,15 @@ export default function Component() {
     setBilans((prev) => prev.filter((b) => b.id !== id));
   };
 
+  const formatDate = (dateString: string) => {
+    const date = new Date(dateString);
+    return date.toLocaleDateString('fr-FR', {
+      day: '2-digit',
+      month: '2-digit',
+      year: 'numeric',
+    });
+  };
+
   // Calculs pour la pagination
   const totalPages = Math.ceil(bilans.length / bilansPerPage);
   const startIndex = (currentPage - 1) * bilansPerPage;
@@ -79,10 +107,42 @@ export default function Component() {
   // État vide - aucun bilan
   if (bilans.length === 0) {
     return (
-      <EmptyState
-        onNewPatient={() => setIsNewPatientModalOpen(true)}
-        onExistingPatient={() => setIsExistingPatientModalOpen(true)}
-      />
+      <div className="min-h-screen bg-gray-50 p-6">
+        <div className="max-w-4xl mx-auto">
+          <Card className="border-2 border-dashed border-gray-200">
+            <CardContent className="flex flex-col items-center justify-center py-12 text-center">
+              <FileText className="w-16 h-16 text-gray-400 mb-6" />
+              <h2 className="text-2xl font-semibold text-gray-900 mb-2">
+                Aucun bilan disponible
+              </h2>
+              <p className="text-gray-600 mb-6 max-w-md">
+                Il semble que vous n&rsquo;ayez pas encore rédigé de bilan.
+                Commencez par en créer un nouveau.
+              </p>
+              <DropdownMenu>
+                <DropdownMenuTrigger asChild>
+                  <Button className="bg-blue-600 hover:bg-blue-700">
+                    <Plus className="w-4 h-4 mr-2" />
+                    Rédiger un nouveau bilan
+                  </Button>
+                </DropdownMenuTrigger>
+                <DropdownMenuContent align="center">
+                  <DropdownMenuItem
+                    onClick={() => setIsNewPatientModalOpen(true)}
+                  >
+                    Nouveau patient
+                  </DropdownMenuItem>
+                  <DropdownMenuItem
+                    onClick={() => setIsExistingPatientModalOpen(true)}
+                  >
+                    Patient existant
+                  </DropdownMenuItem>
+                </DropdownMenuContent>
+              </DropdownMenu>
+            </CardContent>
+          </Card>
+        </div>
+      </div>
     );
   }
 
@@ -133,20 +193,107 @@ export default function Component() {
           {/* Tableau des bilans */}
           <Card>
             <CardContent className="p-6">
-              <BilansTable
-                bilans={currentBilansPage}
-                totalCount={bilans.length}
-                onRowClick={(id) => navigate(`/bilan/${id}`)}
-                onDelete={(b) => setToDelete(b)}
-              />
-              <PaginationControls
-                currentPage={currentPage}
-                totalPages={totalPages}
-                startIndex={startIndex}
-                endIndex={endIndex}
-                totalCount={bilans.length}
-                onPageChange={goToPage}
-              />
+              <div className="flex items-center justify-between mb-6">
+                <h3 className="text-xl font-semibold text-gray-900">
+                  Historique des bilans
+                </h3>
+                <span className="text-sm text-gray-500">
+                  {bilans.length} bilan{bilans.length > 1 ? 's' : ''}
+                </span>
+              </div>
+
+              <div className="overflow-x-auto">
+                <Table>
+                  <TableHeader>
+                    <TableRow>
+                      <TableHead className="w-32">Date</TableHead>
+                      <TableHead>Nom Patient</TableHead>
+                      <TableHead>Nom bilan</TableHead>
+                      <TableHead className="w-10" />
+                    </TableRow>
+                  </TableHeader>
+                  <TableBody>
+                    {currentBilansPage.map((bilan) => (
+                      <TableRow
+                        key={bilan.id}
+                        onClick={() => navigate(`/bilan/${bilan.id}`)}
+                        className="hover:bg-gray-200 cursor-pointer"
+                      >
+                        <TableCell className="font-medium">
+                          {formatDate(bilan.date)}
+                        </TableCell>
+                        <TableCell>
+                          {bilan.patient
+                            ? `${bilan.patient.firstName} ${bilan.patient.lastName}`
+                            : ''}
+                        </TableCell>
+                        <TableCell>{bilan.bilanType?.name ?? ''}</TableCell>
+                        <TableCell onClick={(e) => e.stopPropagation()}>
+                          <Button
+                            size="icon"
+                            variant="ghost"
+                            className="text-red-600"
+                            aria-label="Supprimer le bilan"
+                            onClick={() => setToDelete(bilan)}
+                          >
+                            <Trash2 className="w-4 h-4" />
+                          </Button>
+                        </TableCell>
+                      </TableRow>
+                    ))}
+                  </TableBody>
+                </Table>
+              </div>
+
+              {/* Pagination */}
+              {totalPages > 1 && (
+                <div className="flex items-center justify-between mt-6">
+                  <div className="text-sm text-gray-500">
+                    Affichage de {startIndex + 1} à{' '}
+                    {Math.min(endIndex, bilans.length)} sur {bilans.length}{' '}
+                    bilans
+                  </div>
+                  <div className="flex items-center space-x-2">
+                    <Button
+                      variant="outline"
+                      size="sm"
+                      onClick={() => goToPage(currentPage - 1)}
+                      disabled={currentPage === 1}
+                    >
+                      <ChevronLeft className="w-4 h-4" />
+                      Précédent
+                    </Button>
+
+                    <div className="flex items-center space-x-1">
+                      {Array.from({ length: totalPages }, (_, i) => i + 1).map(
+                        (page) => (
+                          <Button
+                            key={page}
+                            variant={
+                              currentPage === page ? 'default' : 'outline'
+                            }
+                            size="sm"
+                            onClick={() => goToPage(page)}
+                            className="w-8 h-8 p-0"
+                          >
+                            {page}
+                          </Button>
+                        ),
+                      )}
+                    </div>
+
+                    <Button
+                      variant="outline"
+                      size="sm"
+                      onClick={() => goToPage(currentPage + 1)}
+                      disabled={currentPage === totalPages}
+                    >
+                      Suivant
+                      <ChevronRight className="w-4 h-4" />
+                    </Button>
+                  </div>
+                </div>
+              )}
             </CardContent>
           </Card>
         </div>
@@ -161,16 +308,27 @@ export default function Component() {
         onClose={() => setIsExistingPatientModalOpen(false)}
         onPatientSelected={createBilan}
       />
-      <DeleteDialog
-        open={!!toDelete}
-        onCancel={() => setToDelete(null)}
-        onConfirm={async () => {
-          if (toDelete) {
-            await removeBilan(toDelete.id);
-            setToDelete(null);
-          }
-        }}
-      />
+      <AlertDialog open={!!toDelete} onOpenChange={() => setToDelete(null)}>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Supprimer ce bilan ?</AlertDialogTitle>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Annuler</AlertDialogCancel>
+            <AlertDialogAction
+              className="bg-red-600 hover:bg-red-700"
+              onClick={async () => {
+                if (toDelete) {
+                  await removeBilan(toDelete.id);
+                  setToDelete(null);
+                }
+              }}
+            >
+              Supprimer
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </>
   );
 }
